Add share button functionality in TripIternary screen
